Separated batches and deduplicateStagingRecords
