+++ packages/jwt-audit-analytics-writer/src/model/domain/errors.ts
@@ -3,6 +3,7 @@ import { InternalError } from "pagopa-interop-kpi-models";
 export const errorCodes = {
   insertStagingRecordsError: "INSERT_STAGING_RECORDS_ERROR",
   mergeDataError: "MERGE_DATA_ERROR",
+  deduplicateStagingRecordsError: "DEDUPLICATE_STAGING_RECORDS_ERROR",
 } as const;
 
 export type ErrorCodes = keyof typeof errorCodes;
@@ -20,5 +21,14 @@ export function mergeDataError(detail: unknown): InternalError<ErrorCodes> {
   return new InternalError({
     detail: `Database error merging data: ${detail}`,
     code: "mergeDataError",
+  });
+}
+
+export function deduplicateStagingRecordsError(
+  detail: unknown
+): InternalError<ErrorCodes> {
+  return new InternalError({
+    detail: `Database error deduplicating staging records data: ${detail}`,
+    code: "deduplicateStagingRecordsError",
   });
 }
